remove qgemm_common
